+++ src/main/java/com/samcomo/dbz/global/config/CorsMvcConfig.java
@@ -1,0 +1,17 @@
+package com.samcomo.dbz.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class CorsMvcConfig implements WebMvcConfigurer {
+
+  @Override
+  public void addCorsMappings(CorsRegistry corsRegistry){
+
+    corsRegistry.addMapping("/**")
+        .allowedOrigins("http://localhost:5173");
+  }
+
+}
