+++ src/main/java/com/samcomo/dbz/notification/service/NotificationService.java
@@ -1,9 +1,8 @@
 package com.samcomo.dbz.notification.service;
 
-import com.samcomo.dbz.chat.dto.ChatMessageDto;
+import com.samcomo.dbz.chat.model.dto.ChatMessageDto;
 import com.samcomo.dbz.notification.model.dto.NotificationDto;
 import com.samcomo.dbz.report.model.dto.ReportDto;
-import java.io.IOException;
 import java.util.List;
 import org.springframework.stereotype.Service;
