+++ src/main/java/com/samcomo/dbz/notification/service/NotificationServiceImpl.java
@@ -9,7 +9,7 @@ import com.google.firebase.messaging.BatchResponse;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.google.firebase.messaging.MulticastMessage;
-import com.samcomo.dbz.chat.dto.ChatMessageDto;
+import com.samcomo.dbz.chat.model.dto.ChatMessageDto;
 import com.samcomo.dbz.global.exception.ErrorCode;
 import com.samcomo.dbz.member.exception.MemberException;
 import com.samcomo.dbz.member.model.entity.Member;
